backend/requirements.txt
asgiref==3.10.0
Django==5.1.1
django-cors-headers==4.4.0
django-jazzmin==3.0.1
djangorestframework==3.15.2
pillow==10.4.0
sqlparse==0.5.3
tzdata==2025.2
